+++ src/routing/mod.rs
@@ -64,6 +64,7 @@ mod core;
 mod dkg;
 mod ed25519;
 mod error;
+pub(crate) mod log_markers;
 mod messages;
 mod node;
 mod peer;